+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
 ### 0.11.0-beta
-* Add `Precondition.CheckFailed`
+* Added `Precondition.CheckFailed`
+* Added `TableContext.TransactWriteItems`, `TransactWrite` DU, `TransactWriteItemsRequest.TransactionCanceledConditionalCheckFailed`
 
 ### 0.10.1-beta
 * Fixed accidentally removed/renamed legacy factory methods (`TableContext.Create`/`TableContext.CreateAsync`)